(TK-275) Implement "simple" endpoint for plaintext responses
This commit adds two new endpoints to tk-status that support no query
parameters and return only simple plaintext messages and relevant status
codes.

Note that in order to handle the potential for either plaintext or
json responses in our error handling middleware, there is a mix of
compojure and comidi functions when definiing tk-status's ring handler.
We use compojure's functions to compose handlers which have been
generated by comidi and then had middleware applied to them.

We required this because we could no longer apply the same list of
middleware to the entire ring handler (since now the json routes and
plaintext routes required different sets of middleware).